+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (fix) fse-958 | packages/copilot 1.0.11 | Close copilot modal after clicking on Learn More
 - (chore) | packages/constants-helper 1.0.11 | Change Feedback URL
 - (chore) | packages/helpers 1.0.11 | Read proxy URL correctly for WAGMI
 - (chore) | packages/widgets 1.0.1 packages/evmos-wallet 1.0.25 | migrate osmosis and stride outpost to ICS20 precompile